REF stats GUI: move reused components into _wxgui.stats.utils
