+++ eelbrain/_wxgui/load_stcs.py
@@ -10,6 +10,8 @@ import wx
 
 from . import get_app
 from .frame import EelbrainFrame
+from .stats import StatsFrame
+from .stats.utils import FactorPanel, TitleSizer
 from .._io.stc_dataset import DatasetSTCLoader
 
 
@@ -116,29 +118,9 @@ class STCLoaderFrame(EelbrainFrame):
         self.Close()
 
     def OnLaunchStats(self, evt):
-        from .stats import StatsFrame
         ds = self._get_dataset()
         StatsFrame(None, self.loader, ds)
         self.Close()
-
-
-class FactorPanel(wx.Panel):
-    """Panel to display a factor and its level names"""
-    def __init__(self, parent, levels, idx, editable=True):
-        super(FactorPanel, self).__init__(parent)
-        if editable:
-            self.factor_ctl = wx.TextCtrl(self, value="factor_%d" % idx)
-        else:
-            self.factor_ctl = wx.StaticText(self, label="factor_%d" % idx)
-            self.factor_ctl.SetFont(wx.Font.Bold(self.factor_ctl.GetFont()))
-        level_names = ["- " + i for i in levels]
-        level_ctl = wx.StaticText(self)
-        level_ctl.SetLabel("\n".join(level_names))
-        sizer = wx.BoxSizer(wx.VERTICAL)
-        sizer.Add(self.factor_ctl)
-        sizer.Add(level_ctl, 0, wx.EXPAND | wx.TOP, 10)
-        sizer.Layout()
-        self.SetSizer(sizer)
 
 
 class MRIPanel(wx.Panel):
@@ -176,12 +158,3 @@ class MRIPanel(wx.Panel):
         sizer.Add(hs, 0, wx.BOTTOM, 5)
         sizer.Layout()
         self.SetSizer(sizer)
-
-
-class TitleSizer(wx.BoxSizer):
-    def __init__(self, parent, title):
-        super().__init__(wx.HORIZONTAL)
-        self.ctl = wx.StaticText(parent, label=title)
-        font = wx.Font(14, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        self.ctl.SetFont(font)
-        self.Add(self.ctl)
